requirements.txt
pygame>=2.5.0
PyOpenGL>=3.1.7
noise>=1.2.2
